Status: Moved to slatekit.common
